% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suvrel.R, R/suvrel_diag.R
\name{suvrel_diag}
\alias{suvrel_diag}
\alias{suvrel_diag}
\title{Estimate the metric that maximize the intergroup distance while minimizes
the intragroup distance}
\usage{
suvrel_diag(datum, value_var, group, feature)

suvrel_diag(datum, value_var, group, feature)
}
\arguments{
\item{datum}{data.table. Data with one allele per line.}

\item{datum}{data.table. Data with one allele per line.}
}
\value{
data.table metric

data.table metric
}
\description{
\code{suvrel_diag()} Filter SNPs with MAF bellow a threshold

\code{suvrel_diag()} Filter SNPs with MAF bellow a threshold
}
